added knight test and dataset reformatted
