+++ fuseimg/datasets/tests/test_datasets.py
@@ -2,6 +2,7 @@ import os
 import pathlib
 import shutil
 from tempfile import mkdtemp
+from typing import OrderedDict
 import unittest
 from fuse.data.utils.sample import get_sample_id
 from fuse.utils.file_io.file_io import create_dir
@@ -17,6 +18,7 @@ from fuse.eval.evaluator import EvaluatorDefault
 import multiprocessing as mp
 
 from fuseimg.datasets.isic import ISIC
+from fuseimg.datasets.knight import KNIGHT
 from examples.fuse_examples.imaging.classification.isic.golden_members import FULL_GOLDEN_MEMBERS, TEN_GOLDEN_MEMBERS
 
 
@@ -90,6 +92,19 @@ class TestDatasets(unittest.TestCase):
         self.assertEqual(ds[0]["data.input.clinical"].shape[0], 8)
         self.assertTrue(5 in dict(results["metrics.age"]))
 
+    def test_knight(self):
+        data_path = os.environ["KNIGHT_DATA"]
+        split = OrderedDict()
+        split["train"] = [f"case_{id:05d}" for id in range(2)]
+        split["val"] = [f"case_{id:05d}" for id in range(3, 4)]
+        _, _, _, train_ds, valid_ds, _ = KNIGHT.dataset(
+            data_path=data_path,
+            split=split,
+            reset_cache=False,
+            batch_size=1,
+        )
+        self.assertEqual(train_ds[0]["data.input.clinical"].shape[0], 11)
+
     def test_isic(self):
 
         create_dir(self.isic_cache_dir)
@@ -147,4 +162,5 @@ class TestDatasets(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    unittest.main()
+    # unittest.main()
+    TestDatasets.test_knight()
